-- Pioneer의 Weapon 시스템 개선 (ChangeWeapon, Arming, Disarming, SetWeaponType)

+++ Game/Source/Game/Character/Pioneer.h
@@ -243,7 +243,7 @@ public:
 	void FireWeapon(); /** CurrentWeapon을 Fire */
 
 	void SetWeaponType();
-	void ChangeWeapon(float Value); /** Value 값이 1이면 CurrentWeapon의 앞쪽 인덱스, -1이면 CurrentWeapon의 뒤쪽 인덱스 Weapons중 하나로 변경 */
+	void ChangeWeapon(int Value); /** Value 값이 1이면 CurrentWeapon의 앞쪽 인덱스, -1이면 CurrentWeapon의 뒤쪽 인덱스 Weapons중 하나로 변경 */
 
 	void Arming(); /** 비무장 -> 무장(CurrentWeapon) */
 	void Disarming(); /** 무장(CurrentWeapon) -> 비무장 */
